refactor: 볼 카운트 체크하는 함수 분리
